requirements.txt
python-socketio==5.10.0
aiohttp==3.9.1
python-dotenv==1.0.0
pandas==2.1.4
certifi==2023.11.17